feat: tweak DI setup to include memory cache dependency
