+++ LazyCache.AspNetCore/LazyCache.AspNetCore.csproj
@@ -24,6 +24,8 @@
 
   <ItemGroup>
     <PackageReference Include="microsoft.extensions.dependencyinjection.abstractions" Version="2.0.0" />
+    <PackageReference Include="microsoft.extensions.caching.abstractions" Version="2.0.0" />
+    <PackageReference Include="microsoft.extensions.caching.memory" Version="2.0.0" />
     <ProjectReference Include="..\LazyCache\LazyCache.csproj" />
   </ItemGroup>
